warning fixes

+++ tests/helics/apps/combo_tests.cpp
@@ -320,7 +320,7 @@ static void generateFiles_binary(const ghc::filesystem::path& f1, const ghc::fil
     EXPECT_EQ(rec1.messageCount(), 2U);
     EXPECT_EQ(rec1.pointCount(), 3U);
     if (rec1.pointCount() < 3U) {
-        for (int ii = 0; ii < rec1.pointCount(); ++ii) {
+        for (size_t ii = 0; ii < rec1.pointCount(); ++ii) {
             std::cout << "point " << ii << " time=" << std::get<0>(rec1.getValue(ii))
                       << " value= " << std::get<2>(rec1.getValue(ii)) << std::endl;
         }
